chore: add node version to spec file name
Adjust the spec file by adding the node version and env name to the "name" field.

The node version is obtained using the `cargo run -p torus-node -- --version` command. The name is set to "Torus {env} {node_version}". This change makes it easier to know how the spec file was generated.