chore: drop substrate-fixed types (#82)
This is a large one. We've been facing issues with the precision
provided by `substrate-fixed` types for a while now.

This commit drops this dependency in favor of the SDK's provided `Fixed`
types, which handle 18 decimals of precision correctly for most cases.
The change required various tweaks to how we test things. I also
expanded coverage for the emission code.