feat: build and push xtask cli tool to github packages (#86)
This PR adds the `Docker build and push xtask cli tool` action.  
  
When code that changes anything related to the xtask is pushed to the
branch `dev` or `main` this action builds a docker image and publishes
it on github packages.
  
The image tag is `ghcr.io/renlabs-dev/torus-xtask:{commit small sha}`
with the addition of `ghcr.io/renlabs-dev/torus-xtask:latest` if it's
pushed to `main`.

Closes CHAIN-71