feat(permission0): create new curator scope (#97)
Creates a new curator scope for the permission0 system. This change
supersedes the `Curators` storage value in the governance pallet.

A new extrinsic was added to `permission0`: 
```rust
pub fn grant_curator_permission(
	origin: OriginFor<T>,
	grantee: T::AccountId,
	flags: u32,
	cooldown: Option<BlockNumberFor<T>>,
	duration: PermissionDuration<T>,
	revocation: RevocationTerms<T>,
) ;
```

For now, only the sudo (root) key is allowed to grant new permissions,
which we will do through a multisig dispatch. Future versions will allow
curators to re-delegate permissions. The new extrinsic allows setting
different sub-permissions for curator actions:

```rust
        /// Permission to review and process agent applications
        const APPLICATION_REVIEW = 0b0000_0010;
        /// Permission to manage the whitelist (add/remove accounts)
        const WHITELIST_MANAGE   = 0b0000_0100;
        /// Permission to apply penalty factors to agents
        const PENALTY_CONTROL    = 0b0000_1000;
```

The value for `0b1` is reserved for future use and internally assigned
as `ROOT`.

The `flags` field is a bitmask, so OR the different flags together to
get your very own personalized permission.

Finally, I removed the old extrinsics for curators in the governance
pallet.