fix(permission0): allow grantor to delegate emission to itself (#107)
After discussing how the emissions will be initially used, it was clear
to us that there are no cases where it makes sense to delegate emissions
permission to another agent. In this case, the best way to delegate is
to itself, which allows the agent to control how the emissions will flow
and edit the targets at will.