fix(torus0/stake): refund stake before clearing map
This patch fixes the de-registration code to clear the map after the stake has been refunded to the accounts. A previous test case ensure the _wrong_ behavior, which is now fixed.

An agent this week deregistered and got its tokens locked because of order of operations. Luckly, because the TotalStake storage is still wrong, the imbalance represents the amount needed to be refunded without minting new tokens.

Closes CHAIN-103.